requirements.txt
torch
torchvision
torchaudio
diffusers
transformers
accelerate
opencv-python-headless
Pillow
tqdm
numpy